+++ test/test_cadence_logic/MockSystemWrapper.h
@@ -16,7 +16,7 @@ public:
     }
 
     // ISystemWrapper interface
-    unsigned long millis() override {
+    uint32_t millis() override {
         return currentTime;
     }
